requirements.txt
# Matrix Protocol Library with E2EE support
matrix-nio[e2e]>=0.21.0
python-olm>=3.2.0

# XRPL (XRP Ledger) Library
xrpl-py>=2.0.0

# HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.8.0

# JSON Schema Validation
jsonschema>=4.17.0

# Logging
colorlog>=6.7.0

# Date/Time Utilities
python-dateutil>=2.8.2

# Configuration Management
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
